# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateStoreRequest:
    properties:
      banner_url:
        type: string
      category_ids:
        items:
          type: integer
        type: array
      description:
        type: string
      location:
        type: string
      logo_url:
        type: string
      name:
        type: string
      return_policy:
        type: string
      shipping_info:
        type: string
      user_id:
        type: integer
    required:
    - name
    type: object
...
  dto.UpdateStoreRequest:
    properties:
      banner_url:
        type: string
      category_ids:
        items:
          type: integer
        type: array
      description:
        type: string
      is_verified:
        type: boolean
      location:
        type: string
      logo_url:
        type: string
      name:
        type: string
      return_policy:
        type: string
      shipping_info:
        type: string
      status:
        type: string
    type: object
...
  utils.Response:
    properties:
      code:
        type: integer
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@luxe.com
    name: API Support
  description: Production-grade e-commerce backend
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Shopping Platform API
  version: "1.0"
paths:
...
  /admin/stores:
    post:
      consumes:
      - application/json
      description: Create a new store (admin only)
      parameters:
      - description: Store details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateStoreRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.StoreResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Create store
      tags:
      - Stores
  /admin/stores/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a store (admin only)
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Delete store
      tags:
      - Stores
    get:
      consumes:
      - application/json
      description: Fetch store details by primary key for admin edit screens
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.StoreResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get store by ID (admin)
      tags:
      - Stores
    put:
      consumes:
      - application/json
      description: Update store details (admin only)
      parameters:
      - description: Store ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated store data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.StoreResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Update store
      tags:
      - Stores
...
  /stores:
    get:
      consumes:
      - application/json
      description: Get stores with pagination, search, location, rating, category
        filters
      parameters:
      - default: 20
        description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Number of items to skip
        in: query
        name: offset
        type: integer
      - description: Search by name or description
        in: query
        name: search
        type: string
      - description: Filter by location (partial match)
        in: query
        name: location
        type: string
      - description: Minimum rating (0-5)
        in: query
        name: min_rating
        type: number
      - description: Filter by category slug
        in: query
        name: category_slug
        type: string
      - description: Sort order (rating, followers, newest)
        in: query
        name: sort_by
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    limit:
                      type: integer
                    offset:
                      type: integer
                    stores:
                      items:
                        $ref: '#/definitions/dto.StoreResponse'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: List stores
      tags:
      - Stores
  /stores/{slug}:
    get:
      consumes:
      - application/json
      description: Fetch store details including categories and stats
      parameters:
      - description: Store slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.StoreResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get store by slug
      tags:
      - Stores
  /stores/{slug}/follow:
    delete:
      consumes:
      - application/json
      description: Unfollow a store (authenticated)
      parameters:
      - description: Store slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Unfollow a store
      tags:
      - Stores
    post:
      consumes:
      - application/json
      description: Follow a store (authenticated)
      parameters:
      - description: Store slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Follow a store
      tags:
      - Stores
  /stores/{slug}/products:
    get:
      consumes:
      - application/json
      description: List products belonging to a store with product filters & pagination
      parameters:
      - description: Store slug
        in: path
        name: slug
        required: true
        type: string
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Number of items to skip
        in: query
        name: offset
        type: integer
      - description: Search by product name (partial match)
        in: query
        name: name
        type: string
      - description: Filter by category ID
        in: query
        name: category_id
        type: integer
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      - description: Minimum rating
        in: query
        name: min_rating
        type: number
      - description: Digital products only
        in: query
        name: is_digital
        type: boolean
      - description: New arrivals only
        in: query
        name: is_new
        type: boolean
      - description: Sort order (price_asc,price_desc,rating_desc,newest)
        in: query
        name: sort
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  properties:
                    limit:
                      type: integer
                    offset:
                      type: integer
                    products:
                      items:
                        $ref: '#/definitions/dto.ProductResponse'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      summary: Get products of a store
      tags:
      - Stores
...
  /stores/{slug}/reviews/me:
    get:
      description: Returns the authenticated user's review for a store, if any
      parameters:
      - description: Store slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: Get my store review
      tags:
      - Stores
...
  /vendor/stores:
    get:
      consumes:
      - application/json
      description: Returns stores owned by the current seller, or all stores for admins/moderators
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.StoreResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: List vendor stores
      tags:
      - Vendor
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"